+++ app_estudo/domain/__init__.py
@@ -1,6 +1,13 @@
 """Camada de dominio do projeto."""
 
 from .listening_evaluation import ListeningEvaluation, evaluate_listening_item
+from .source_media import SourceMedia, SourceMetadata
 from .study_item import CuratedStudyItem
 
-__all__ = ["CuratedStudyItem", "ListeningEvaluation", "evaluate_listening_item"]
+__all__ = [
+	"CuratedStudyItem",
+	"ListeningEvaluation",
+	"SourceMedia",
+	"SourceMetadata",
+	"evaluate_listening_item",
+]